Initial commit of multiloader dev environment. Forge client would not run while code remained in its folder, so it has been stripped down and will need to be added back.
Problems downloading sources for class files have been encountered. I will also need to figure out if the empty main and test modules repeatedly adding themselves to the project (the ones outside the common, forge, & fabric folders) are intended to be there or are a result of a hiccup during migration to a multiloader environment.